+++ src/rpc/methods/msig.rs
@@ -3,7 +3,7 @@
 
 use crate::rpc::error::ServerError;
 use crate::rpc::types::ApiTipsetKey;
-use crate::rpc::types::MsigVesting;
+use crate::rpc::types::*;
 use crate::rpc::{ApiVersion, Ctx, Permission, RpcMethod};
 use crate::shim::{address::Address, econ::TokenAmount};
 use fil_actor_interface::multisig;
@@ -12,11 +12,77 @@ use num_bigint::BigInt;
 
 macro_rules! for_each_method {
     ($callback:ident) => {
+        $callback!(crate::rpc::msig::MsigGetAvailableBalance);
+        $callback!(crate::rpc::msig::MsigGetPending);
         $callback!(crate::rpc::msig::MsigGetVested);
         $callback!(crate::rpc::msig::MsigGetVestingSchedule);
     };
 }
 pub(crate) use for_each_method;
+
+pub enum MsigGetAvailableBalance {}
+
+impl RpcMethod<2> for MsigGetAvailableBalance {
+    const NAME: &'static str = "Filecoin.MsigGetAvailableBalance";
+    const PARAM_NAMES: [&'static str; 2] = ["address", "tipset_key"];
+    const API_VERSION: ApiVersion = ApiVersion::V0;
+    const PERMISSION: Permission = Permission::Read;
+
+    type Params = (Address, ApiTipsetKey);
+    type Ok = TokenAmount;
+
+    async fn handle(
+        ctx: Ctx<impl Blockstore>,
+        (address, ApiTipsetKey(tsk)): Self::Params,
+    ) -> Result<Self::Ok, ServerError> {
+        let ts = ctx.chain_store.load_required_tipset_or_heaviest(&tsk)?;
+        let height = ts.epoch();
+        let actor = ctx
+            .state_manager
+            .get_required_actor(&address, *ts.parent_state())?;
+        let actor_balance = TokenAmount::from(&actor.balance);
+        let ms = multisig::State::load(ctx.store(), actor.code, actor.state)?;
+        let locked_balance = ms.locked_balance(height)?.into();
+        let avail_balance = &actor_balance - locked_balance;
+        Ok(avail_balance)
+    }
+}
+
+pub enum MsigGetPending {}
+
+impl RpcMethod<2> for MsigGetPending {
+    const NAME: &'static str = "Filecoin.MsigGetPending";
+    const PARAM_NAMES: [&'static str; 2] = ["address", "tipset_key"];
+    const API_VERSION: ApiVersion = ApiVersion::V0;
+    const PERMISSION: Permission = Permission::Read;
+
+    type Params = (Address, ApiTipsetKey);
+    type Ok = Vec<Transaction>;
+
+    async fn handle(
+        ctx: Ctx<impl Blockstore>,
+        (address, ApiTipsetKey(tsk)): Self::Params,
+    ) -> Result<Self::Ok, ServerError> {
+        let ts = ctx.chain_store.load_required_tipset_or_heaviest(&tsk)?;
+        let actor = ctx
+            .state_manager
+            .get_required_actor(&address, *ts.parent_state())?;
+        let ms = multisig::State::load(ctx.store(), actor.code, actor.state)?;
+        let txns = ms
+            .get_pending_txn(ctx.store())?
+            .iter()
+            .map(|txn| Transaction {
+                id: txn.id,
+                to: txn.to.into(),
+                value: txn.value.clone().into(),
+                method: txn.method,
+                params: txn.params.clone(),
+                approved: txn.approved.iter().map(|item| item.into()).collect(),
+            })
+            .collect();
+        Ok(txns)
+    }
+}
 
 pub enum MsigGetVested {}
 impl RpcMethod<3> for MsigGetVested {
